+++ app/controllers/products_controller.rb
@@ -17,4 +17,9 @@ class ProductsController < ApplicationController
   def show
 
   end
+
+  private
+  def product_params
+    params.require(:product).permit(images: [])
+  end
 end