+++ app/models/product.rb
@@ -1,3 +1,5 @@
 class Product < ApplicationRecord
-  belongs_to :product_group
+  belongs_to :product_group, optional: true
+  belongs_to :producer, optional: true
+  has_many_attached :images
 end